update transaction cashflow

+++ app/Http/Controllers/CashInController.php
@@ -7,6 +7,8 @@ use App\Models\Fund;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
+use Throwable;
 
 class CashInController extends Controller
 {
@@ -15,7 +17,7 @@ class CashInController extends Controller
      */
     public function index()
     {
-        $data = CashIn::paginate(5);
+        $data = CashIn::where('company_id', Auth::user()->company_id)->paginate(5);
         return view('dashboard.finance.cash-in.index', ['data' => $data]);
     }
 
@@ -33,24 +35,27 @@ class CashInController extends Controller
      */
     public function store(Request $request)
     {
-        $validate = $request->validate([
-            'fund' => 'required',
-            'remark' => 'required',
-            'datetime' => 'required',
-            'type' => 'required',
-        ]);
+        $query_data = array();
+        try {
+            DB::beginTransaction();
 
-        $store = CashIn::create([
-            'company_id' => Auth::user()->company_id,
-            'fund' => $validate['fund'],
-            'remark' => $validate['remark'],
-            'datetime' => $validate['datetime'],
-            'type' => $validate['type'],
-        ]);
+            $validate = $request->validate([
+                'fund' => 'required',
+                'remark' => 'required',
+                'datetime' => 'required',
+                'type' => 'required',
+            ]);
 
-        $query_data = ['periode' => Carbon::parse($validate['datetime'])->format('Y-m-d')];
+            $store = CashIn::create([
+                'company_id' => Auth::user()->company_id,
+                'fund' => $validate['fund'],
+                'remark' => $validate['remark'],
+                'datetime' => $validate['datetime'],
+                'type' => $validate['type'],
+            ]);
 
-        if ($store) {
+            $query_data = ['periode' => Carbon::parse($validate['datetime'])->format('Y-m-d')];
+
             $fund = Fund::where(
                 "company_id",
                 Auth::user()->company_id
@@ -63,8 +68,10 @@ class CashInController extends Controller
                 "fund" => $fund->fund + $validate["fund"]
             ]);
 
+            DB::commit();
             return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to create cash in");
-        } else {
+        } catch (Throwable $error) {
+            DB::rollBack();
             return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to create cash in");
         }
     }
@@ -77,7 +84,7 @@ class CashInController extends Controller
         $data = CashIn::findOrFail($id);
         $funds = Fund::where('company_id', Auth::user()->company_id)->get();
 
-        if ($data->company_id == Auth::user()->company_id) {
+        if ($data && $data->company_id == Auth::user()->company_id) {
             return view('dashboard.finance.cash-in.edit', ["data" => $data, 'funds' => $funds]);
         } else {
             return view('dashboard.finance.cash-in')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
@@ -89,29 +96,31 @@ class CashInController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $validate = $request->validate([
-            'fund' => 'required',
-            'remark' => 'required',
-            'datetime' => 'required',
-            'type' => 'required',
-        ]);
-
-        $data = CashIn::findOrFail($id);
-        $type = $data->type;
-        $amount = $data->fund;
-
-        if ($data->company_id == Auth::user()->company_id) {
-            $update = $data->update([
-                'company_id' => Auth::user()->company_id,
-                'fund' => $validate['fund'],
-                'remark' => $validate['remark'],
-                'datetime' => $validate['datetime'],
-                'type' => $validate['type'],
+        $query_data = array();
+        try {
+            DB::beginTransaction();
+            $validate = $request->validate([
+                'fund' => 'required',
+                'remark' => 'required',
+                'datetime' => 'required',
+                'type' => 'required',
             ]);
-
-            $query_data = ['periode' => Carbon::parse($validate['datetime'])->format('Y-m-d')];
-
-            if ($update) {
+    
+            $data = CashIn::findOrFail($id);
+            $type = $data->type;
+            $amount = $data->fund;
+    
+            if ($data && $data->company_id == Auth::user()->company_id) {
+                $update = $data->update([
+                    'company_id' => Auth::user()->company_id,
+                    'fund' => $validate['fund'],
+                    'remark' => $validate['remark'],
+                    'datetime' => $validate['datetime'],
+                    'type' => $validate['type'],
+                ]);
+    
+                $query_data = ['periode' => Carbon::parse($validate['datetime'])->format('Y-m-d')];
+    
                 if ($type == $validate["type"]) {
                     $fund = Fund::where(
                         "company_id",
@@ -134,20 +143,23 @@ class CashInController extends Controller
                         Auth::user()->company_id
                     )->where('type', $validate['type'])->first();
 
-                    $fundOld->where('type', $type)->update([
+                    $fundOld->update([
                         "fund" => $fundOld->fund - $amount
                     ]);
 
-                    $fund->where('type', $validate['type'])->update([
+                    $fund->update([
                         "fund" => $fund->fund + $validate["fund"]
                     ]);
                 }
+                DB::commit();
                 return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to update cash in");
             } else {
-                return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to update cash in");
+                DB::rollBack();
+                return abort(404);
             }
-        } else {
-            return abort(404);
+        } catch (Throwable $error) {
+            DB::rollBack();
+            return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to update cash in");
         }
     }
 
@@ -157,20 +169,32 @@ class CashInController extends Controller
     public function destroy(string $id, Request $request)
     {
         $query_data = array();
-        if ($request->periode) {
-            $query_data = ['periode' => $request->periode];
-        }
-
-        $data = CashIn::findOrFail($id);
-        if ($data->company_id == Auth::user()->company_id) {
-            $delete =  CashIn::destroy($id);
-            if ($delete) {
+        try {
+            DB::beginTransaction();
+            if ($request->periode) {
+                $query_data = ['periode' => $request->periode];
+            }
+    
+            $data = CashIn::findOrFail($id);
+            if ($data && $data->company_id == Auth::user()->company_id) {
+                $fund = Fund::where(
+                    "company_id",
+                    Auth::user()->company_id
+                )->where('type', $data->type)->first();
+                $fund->update([
+                    'fund' => $fund->fund - $data->fund,
+                ]);
+                $delete =  CashIn::destroy($id);
+                
+                DB::commit();
                 return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to delete cash in");
             } else {
-                return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to delete cash in");
+                DB::rollBack();
+                return abort(404);
             }
-        } else {
-            return abort(404);
+        } catch (Throwable $error) {
+            DB::rollBack();
+            return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to delete cash in");
         }
     }
 }
